+++ apps/desktop/lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 36
+/// Strings: 48
 ///
-/// Built on 2026-09-13 at 17:53 UTC
+/// Built on 2026-09-14 at 22:16 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import